Frontend. Refactor Place creation and editing forms to use PlaceForm component (#26)
