+++ fotolyrik_frontend/components/forms/PlaceForm.vue
@@ -1,8 +1,4 @@
 <script setup lang="ts">
-import type { Place } from "~/utils/types";
-import { onMounted, ref } from "vue";
-import { useToast } from "primevue/usetoast";
-import { getNode } from "@formkit/core";
 import maplibregl from "maplibre-gl";
 import "maplibre-gl/dist/maplibre-gl.css";
 
@@ -12,172 +8,210 @@ const props = defineProps<{
   place?: Place;
 }>();
 
+const placeStore = usePlaceStore();
 const toast = useToast();
-const submitted = ref(false);
-const place_store = usePlaceStore();
 
-type PlaceInput = Omit<Place, 'id' | 'createdBy' | 'createdDate' | 'lastModifiedBy' | 'lastModifiedDate'>;
+const formRef = ref<any>(null);
 
-const submit = async (formData: Partial<PlaceInput>) => {
-  try {
-    if (props.action === 'create') {
-      await place_store.createPlace(formData);
-      submitted.value = true;
-      toast.add({severity: 'success', summary: 'Erfolg', detail: 'Erfolgreich erstellt', life: 3000});
-      const form = getNode('place_creation');
-      form?.reset();
-      if (current_marker) {
-        current_marker.remove();
-        current_marker = null;
+let map: maplibregl.Map;
+let marker: maplibregl.Marker | null;
+
+const lat = ref<number | null>(null);
+const lng = ref<number | null>(null);
+
+const onFormSubmit = async (e: any) => {
+  if (e.valid) {
+    try {
+      if (props.action === 'create') {
+        await placeStore.createPlace(e.values);
+        toast.add({severity: 'success', summary: 'Erfolg', detail: 'Erfolgreich erstellt', life: 3000});
+        navigateTo('/places')
+      } else if (props.action === 'edit' && props.place?.id) {
+        await placeStore.updatePlace(e.values, props.place.id);
+        toast.add({severity: 'success', summary: 'Erfolg', detail: 'Erfolgreich aktualisiert', life: 3000});
+        navigateTo(`/places/${props.place?.id}`);
       }
-    } else if (props.action === 'edit' && props.place?.id) {
-      await place_store.updatePlace(formData, props.place.id);
-      submitted.value = true;
-      toast.add({severity: 'success', summary: 'Erfolg', detail: 'Erfolgreich upgedated', life: 3000});
-      navigateTo(`/photopoems/${props.place?.id}`);
+    } catch (error) {
+      console.log(error);
+      toast.add({severity: 'error', summary: 'Fehler', detail: 'Ein Fehler ist aufgetreten', life: 3000});
     }
-  } catch (error) {
-    console.log(error)
-    toast.add({
-      severity: 'error',
-      summary: 'Fehler',
-      detail: 'Fehler beim Erstellen des Ort-Objektes',
-      life: 3000
-    });
   }
 };
 
-let current_marker: maplibregl.Marker | null = null;
-const latitude = ref<number|null>(null);
-const longitude = ref<number|null>(null);
+const onCoordinatesUpdate = () => {
+  if (lat.value == null || lng.value == null) return;
+  const coords: [number, number] = [lng.value, lat.value];
+  if (marker) {
+    marker.setLngLat(coords);
+  } else {
+    marker = new maplibregl.Marker({ color: "#063D79" })
+        .setLngLat(coords)
+        .addTo(map);
+  }
+  map.setCenter(coords);
+  map.setZoom(7);
+};
 
 onMounted(async () => {
-  const map = new maplibregl.Map({
+  if (props.place?.latitude && props.place?.longitude) {
+    lat.value = props.place.latitude;
+    lng.value = props.place.longitude;
+  }
+  map = new maplibregl.Map({
     container: "map",
-    zoom: 4.5,
-    center: [11, 51],
     style: {
       version: 8,
-      glyphs: "https://demotiles.maplibre.org/font/{fontstack}/{range}.pbf",
       sources: {
         osm: {
-          type: "raster",
-          tiles: ["https://tile.openstreetmap.de/{z}/{x}/{y}.png"],
+          type: 'raster',
+          tiles: ['https://tile.openstreetmap.de/{z}/{x}/{y}.png'],
           tileSize: 256,
-          attribution: "&copy; OpenStreetMap Contributors",
-        },
+          attribution: '&copy; OpenStreetMap Contributors'
+        }
       },
       layers: [
         {
-          id: "osm-layer",
-          type: "raster",
-          source: "osm",
-        },
-      ],
+          id:'osm-layer',
+          type: 'raster',
+          source: 'osm'
+        }
+      ]
     },
+    center: [10.4515, 51.1657],
+    zoom: 5
   });
-
-  map.on('click', (e) => {
-    let { lng, lat } = e.lngLat
-    lng = Number(lng.toFixed(5))
-    lat = Number(lat.toFixed(5))
-
-    if (current_marker) {
-      current_marker.remove();
-    }
-    current_marker = new maplibregl.Marker()
-        .setLngLat([lng, lat])
-        .setPopup(new maplibregl.Popup())
+  map.addControl(new maplibregl.NavigationControl(), 'top-right');
+  if (lat.value != null && lng.value != null) {
+    const coords: [number, number] = [lng.value, lat.value];
+    marker = new maplibregl.Marker({ color: "#063D79" })
+        .setLngLat(coords)
         .addTo(map);
-
-    latitude.value = lat
-    longitude.value = lng
-  });
-  map.on('contextmenu', () => {
-    if (current_marker) {
-      current_marker.remove();
-      current_marker = null;
+    map.setCenter(coords);
+    map.setZoom(7);
+  }
+  map.on('click', (e) => {
+    lat.value = Number(e.lngLat.lat.toFixed(6));
+    lng.value = Number(e.lngLat.lng.toFixed(6));
+    formRef.value.setFieldValue('latitude', lat.value);
+    formRef.value.setFieldValue('longitude', lng.value);
+    if (marker) {
+      marker.setLngLat(e.lngLat);
+    } else {
+      marker = new maplibregl.Marker({ color: "#063D79" })
+          .setLngLat(e.lngLat)
+          .addTo(map);
     }
-    latitude.value = null;
-    longitude.value = null;
   });
+  map.on('contextmenu', (e) => {
+    lat.value = null;
+    lng.value = null;
+    formRef.value.setFieldValue('latitude', null);
+    formRef.value.setFieldValue('longitude', null);
+    if (marker) {
+      marker.remove();
+      marker = null;
+    }
+  })
 });
 </script>
 
 <template>
-  <div class="flex flex-col gap-2">
-    <h1 class="text-2xl outfit-headline font-bold">{{ props.header }}</h1>
-    <div>
-      Klicken Sie auf die Karte, um einen Ort zu platzieren. Der Längen- und Breitengrad werden automatisch in die entsprechenden Formularfelder übernommen.
-    </div>
-    <div>
-      <div id="map" class="h-[400px] w-full"></div>
-    </div>
-    <h1 class="text-2xl outfit-headline font-bold text-[#063D79]">Neuen Ort erstellen</h1>
-    <p class="roboto-plain">Füllen Sie bitte die untenstehenden Felder aus, um einen Ort zu erstellen</p>
-    <FormKit
-        type="form"
-        id="place_creation"
-        :form-class="submitted ? 'hide' : 'show'"
-        submit-label="Erstellen"
-        @submit="submit"
-        #default="{ value }"
-        :actions="false"
-    >
-      <div class="flex flex-col gap-2 border-2 border-solid rounded-md p-5 bg-[#F1F2F2]">
-        <FormKit
-            type="text"
-            name="name"
-            label="Ortsname"
-            placeholder="Berlin"
-            validation="required"
-            prefix-icon="text"
-            outer-class="max-w-full"
-        />
-        <FormKit
-            type="textarea"
-            name="description"
-            label="Beschreibung"
-            placeholder="Hauptstadt Deutschlands"
-            prefix-icon="textarea"
-            outer-class="max-w-full min-w-[0%]"
-        />
-        <div class="flex flex-row space-x-5">
-          <FormKit
-              type="number"
-              number
-              name="latitude"
-              v-model="latitude"
-              label="Breitengrad"
-              placeholder="52.5162"
-              prefix-icon="number"
-              outer-class="max-w-full"
-          />
-          <FormKit
-              type="number"
-              number
-              name="longitude"
-              v-model="longitude"
-              label="Längengrad"
-              placeholder="13.3777"
-              prefix-icon="number"
-              outer-class="max-w-full"
-          />
-        </div>
-        <div class="border-solid border-2 rounded-md p-5 bg-[#F1F2F5] mb-2">
-          <div class="font-mono">JSON-Preview</div>
-          <hr>
-          <pre wrap class="text-sm md:text-base">{{ value }}</pre>
-        </div>
-        <FormKit
-            type="submit"
-            label="Erstellen"
-        />
+  <div class="flex flex-col mx-auto w-[70%] gap-4">
+    <h1 class="text-2xl outfit-headline text-[#063D79] font-bold">{{ props.header }}</h1>
+    <Fieldset legend="Anleitung" class="border-2 border-solid rounded-md overflow-auto">
+      <div class="flex flex-col gap-1 p-2 rounded-md roboto-plain text-black">
+        <p>Füllen Sie bitte die untenstehenden Felder aus, um einen Ort zu erstellen oder anzupassen.</p>
+        <p>Sie können die Koordinaten entweder manuell einfügen oder einen Marker <i class="pi pi-map-marker text-[#063D79]"/> auf der Karte setzen.</p>
+        <p>Falls Sie den Marker löschen wollen, drücken Sie bitte die rechte Maustaste.</p>
       </div>
-    </FormKit>
+    </Fieldset>
+    <div class="flex flex-col gap-2 border-2 border-solid rounded-md p-5 bg-none">
+      <Form
+          ref="formRef"
+          v-slot="$form"
+          @submit="onFormSubmit"
+          class="flex flex-col gap-4"
+          :initial-values="props.place ? props.place : {} as Place"
+      >
+        <FormField v-slot="$field" name="name" class="flex flex-col gap-1 flex-auto">
+          <label for="name" class="font-bold">Ortsname*</label>
+          <IconField>
+            <InputIcon class="pi pi-map-marker" />
+            <InputText
+                id="name"
+                placeholder="Berlin"
+                v-on:keydown.enter.prevent
+                fluid
+            />
+          </IconField>
+          <Message v-if="$form.name?.invalid" severity="error" size="small" variant="simple">
+            {{ $form.name.error.message }}
+          </Message>
+        </FormField>
+        <FormField v-slot="$field" name="description" class="flex flex-col gap-1 flex-auto">
+          <label for="description" class="font-bold">Beschreibung</label>
+          <Textarea
+              id="description"
+              placeholder="Hauptstadt Deutschlands"
+              rows="2"
+              autoResize
+              fluid
+          />
+          <Message v-if="$form.description?.invalid" severity="error" size="small" variant="simple">
+            {{ $form.description.error.message }}
+          </Message>
+        </FormField>
+        <div class="flex flex-col gap-1 flex-auto">
+          <label for="map" class="font-bold">Karte</label>
+          <div id="map" class="h-[600px] w-full rounded-md"/>
+        </div>
+        <div class="flex flex-row space-x-3">
+          <FormField v-slot="$field" name="latitude" class="flex flex-col gap-1 flex-1">
+            <label for="latitude" class="font-bold">Breitengrad</label>
+            <IconField>
+              <InputIcon class="pi pi-map" />
+              <InputNumber
+                  id="latitude"
+                  v-model="lat"
+                  placeholder="52.5162"
+                  :min="-90"
+                  :max="90"
+                  :useGrouping="false"
+                  :minFractionDigits="0"
+                  :maxFractionDigits="6"
+                  v-on:keydown.enter.prevent
+                  @update:modelValue="onCoordinatesUpdate"
+                  fluid
+              />
+            </IconField>
+          </FormField>
+          <FormField v-slot="$field" name="longitude" class="flex flex-col gap-1 flex-1">
+            <label for="longitude" class="font-bold">Längengrad</label>
+            <IconField>
+              <InputIcon class="pi pi-map" />
+              <InputNumber
+                  id="longitude"
+                  v-model="lng"
+                  placeholder="13.3777"
+                  :min="-180"
+                  :max="180"
+                  :useGrouping="false"
+                  :minFractionDigits="0"
+                  :maxFractionDigits="6"
+                  v-on:keydown.enter.prevent
+                  @update:modelValue="onCoordinatesUpdate"
+                  fluid
+              />
+            </IconField>
+          </FormField>
+        </div>
+        <!--
+        <Fieldset legend="Form States" class="h-80 overflow-auto">
+          <pre class="whitespace-pre-wrap">{{ $form }}</pre>
+        </Fieldset>
+        -->
+        <Button type="submit" severity="primary" :label="props.action === 'create' ? 'Erstellen' : 'Ändern'"/>
+      </Form>
+    </div>
   </div>
 </template>
-
-<style scoped>
-
-</style>
